Move USART over to new register interface.
Some of the complex register settings, combined with the encapsulation
of set_mode within a function, called for extending the Register
interface. The core of it is solid, but we're going to want to add more
operators (e.g.  |) and casts over time. Right now I added a to_u32()
function (as well as to_u16() and to_u8() for such sized registers) so
that you can convert a series of type-safe field settings into a raw u32.